+++ assets/src/apps/authoring/components/RightMenu/RightMenu.tsx
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/ban-types */
+import { UiSchema } from '@rjsf/core';
 import { JSONSchema7 } from 'json-schema';
 import { debounce, isEqual } from 'lodash';
 import React, { useCallback, useEffect, useState } from 'react';
@@ -37,6 +38,11 @@ import bankSchema, {
   transformBankModeltoSchema,
   transformBankSchematoModel,
 } from '../PropertyEditor/schemas/bank';
+import bankPropsSchema, {
+  BankPropsUiSchema,
+  transformBankPropsModeltoSchema,
+  transformBankPropsSchematoModel,
+} from '../PropertyEditor/schemas/bankScreen';
 import lessonSchema, {
   lessonUiSchema,
   transformModelToSchema as transformLessonModel,
@@ -79,6 +85,7 @@ const RightMenu: React.FC<any> = () => {
 
   const [scrData, setScreenData] = useState();
   const [scrSchema, setScreenSchema] = useState<JSONSchema7>();
+  const [scrUiSchema, setScreenUiSchema] = useState<UiSchema>();
   const [questionBankData, setBankData] = useState<any>();
   const [questionBankSchema, setBankSchema] = useState<JSONSchema7>();
   const [showConfirmDelete, setShowConfirmDelete] = useState<boolean>(false);
@@ -87,8 +94,13 @@ const RightMenu: React.FC<any> = () => {
       return;
     }
     console.log('CURRENT', { currentActivity, currentLesson });
-    setScreenData(transformScreenModeltoSchema(currentActivity));
-    setScreenSchema(screenSchema);
+    setScreenData(
+      currentSequence?.custom.isBank
+        ? transformBankPropsModeltoSchema(currentActivity)
+        : transformScreenModeltoSchema(currentActivity),
+    );
+    setScreenSchema(currentSequence?.custom.isBank ? bankPropsSchema : screenSchema);
+    setScreenUiSchema(currentSequence?.custom.isBank ? BankPropsUiSchema : screenUiSchema);
 
     setBankData(transformBankModeltoSchema(currentSequence as SequenceEntry<SequenceBank>));
     setBankSchema(bankSchema);
@@ -146,7 +158,9 @@ const RightMenu: React.FC<any> = () => {
   const screenPropertyChangeHandler = useCallback(
     (properties: object) => {
       if (currentActivity) {
-        const modelChanges = transformScreenSchematoModel(properties);
+        const modelChanges = currentSequence?.custom.isBank
+          ? transformBankPropsSchematoModel(properties)
+          : transformScreenSchematoModel(properties);
         console.log('Screen Property Change...', { properties, modelChanges });
         const { title, ...screenModelChanges } = modelChanges;
         const screenChanges = {
@@ -158,7 +172,9 @@ const RightMenu: React.FC<any> = () => {
         if (title) {
           cloneActivity.title = title;
         }
-        debounceSaveScreenSettings(cloneActivity, currentActivity, currentGroup);
+        if (JSON.stringify(cloneActivity) !== JSON.stringify(currentActivity)) {
+          debounceSaveScreenSettings(cloneActivity, currentActivity, currentGroup);
+        }
       }
     },
     [currentActivity],
@@ -197,33 +213,38 @@ const RightMenu: React.FC<any> = () => {
     [],
   );
 
-  const lessonPropertyChangeHandler = (properties: object) => {
-    const modelChanges = transformLessonSchema(properties);
+  const lessonPropertyChangeHandler = useCallback(
+    (properties: object) => {
+      const modelChanges = transformLessonSchema(properties);
 
-    // special consideration for legacy stylesheets
-    if (modelChanges.additionalStylesheets[0] === null) {
-      modelChanges.additionalStylesheets[0] = (currentLesson.additionalStylesheets || [])[0];
-    }
+      // special consideration for legacy stylesheets
+      if (modelChanges.additionalStylesheets[0] === null) {
+        modelChanges.additionalStylesheets[0] = (currentLesson.additionalStylesheets || [])[0];
+      }
 
-    const lessonChanges = {
-      ...currentLesson,
-      ...modelChanges,
-      custom: { ...currentLesson.custom, ...modelChanges.custom },
-    };
-    //need to remove the allowNavigation property
-    //making sure the enableHistory is present before removing that.
-    if (
-      lessonChanges.custom.enableHistory !== undefined &&
-      lessonChanges.custom.allowNavigation !== undefined
-    ) {
-      delete lessonChanges.custom.allowNavigation;
-    }
-    console.log('LESSON PROP CHANGED', { modelChanges, lessonChanges, properties });
+      const lessonChanges = {
+        ...currentLesson,
+        ...modelChanges,
+        custom: { ...currentLesson.custom, ...modelChanges.custom },
+      };
+      //need to remove the allowNavigation property
+      //making sure the enableHistory is present before removing that.
+      if (
+        lessonChanges.custom.enableHistory !== undefined &&
+        lessonChanges.custom.allowNavigation !== undefined
+      ) {
+        delete lessonChanges.custom.allowNavigation;
+      }
+      console.log('LESSON PROP CHANGED', { modelChanges, lessonChanges, properties });
 
-    // need to put a healthy debounce in here, this fires every keystroke
-    // save the page
-    debounceSavePage(lessonChanges);
-  };
+      // need to put a healthy debounce in here, this fires every keystroke
+      // save the page
+      if (JSON.stringify(lessonChanges) !== JSON.stringify(currentLesson)) {
+        debounceSavePage(lessonChanges);
+      }
+    },
+    [currentLesson],
+  );
 
   const debouncePartPropertyChanges = useCallback(
     debounce(
@@ -264,7 +285,7 @@ const RightMenu: React.FC<any> = () => {
         }
         ogPart.custom = modelChanges.custom;
 
-        if (!isEqual(cloneActivity, origActivity)) {
+        if (JSON.stringify(cloneActivity) !== JSON.stringify(origActivity)) {
           dispatch(saveActivity({ activity: cloneActivity }));
         }
       },
@@ -441,8 +462,8 @@ const RightMenu: React.FC<any> = () => {
           {currentActivity && scrData ? (
             <PropertyEditor
               key={currentActivity.id}
-              schema={screenSchema as JSONSchema7}
-              uiSchema={screenUiSchema}
+              schema={scrSchema as JSONSchema7}
+              uiSchema={scrUiSchema as UiSchema}
               value={scrData}
               onChangeHandler={screenPropertyChangeHandler}
             />