+++ assets/src/components/parts/janus-text-flow/Markup.tsx
@@ -105,6 +105,10 @@ const Markup: React.FC<any> = ({
     // add a non breaking space instead of nothing
 
     processedText = processedText.length < 2 ? '\u00a0' : processedText.replace(/ \s/g, '\u00a0 ');
+  } else if (processedText.length !== processedText.trim().length) {
+    // check if text has leading and trailing spaces.
+    //handling the leading blank spacecs in the span
+    processedText = processedText.replace(/\s /g, ' \u00a0');
   }
 
   // this is to support "legacy" SmartSparrow lessons