+++ assets/test/activities/hints_test.ts
@@ -1,6 +1,6 @@
-import { HintActions } from 'components/activities/common/hints/authoring/hintActions';
-import { getHints } from 'components/activities/common/hints/authoring/hintUtils';
+import { DEFAULT_PART_ID } from 'components/activities/common/utils';
 import { makeHint, ScoringStrategy } from 'components/activities/types';
+import { Hints } from 'data/activities/model/hints';
 import { dispatch } from 'utils/test_utils';
 
 describe('authoring hints', () => {
@@ -8,7 +8,7 @@ describe('authoring hints', () => {
     authoring: {
       parts: [
         {
-          id: '1',
+          id: DEFAULT_PART_ID,
           hints: [makeHint(''), makeHint(''), makeHint('')],
           responses: [],
           scoringStrategy: {} as ScoringStrategy,
@@ -18,23 +18,29 @@ describe('authoring hints', () => {
   };
 
   it('can add a cognitive hint before the end of the array', () => {
-    expect(getHints(dispatch(model, HintActions.addHint(makeHint('')))).length).toBeGreaterThan(
-      getHints(model).length,
-    );
+    expect(
+      Hints.byPart(
+        dispatch(model, Hints.addCognitiveHint(makeHint(''), DEFAULT_PART_ID)),
+        DEFAULT_PART_ID,
+      ).length,
+    ).toBeGreaterThan(Hints.byPart(model, DEFAULT_PART_ID).length);
   });
 
   it('can edit a hint', () => {
     const newHintContent = makeHint('new content').content;
-    const firstHint = getHints(model)[0];
+    const firstHint = Hints.byPart(model, DEFAULT_PART_ID)[0];
     expect(
-      getHints(dispatch(model, HintActions.editHint(firstHint.id, newHintContent)))[0],
+      Hints.byPart(
+        dispatch(model, Hints.setContent(firstHint.id, newHintContent)),
+        DEFAULT_PART_ID,
+      )[0],
     ).toHaveProperty('content', newHintContent);
   });
 
   it('can remove a hint', () => {
-    const firstHint = getHints(model)[0];
+    const firstHint = Hints.byPart(model, DEFAULT_PART_ID)[0];
     expect(
-      getHints(dispatch(model, HintActions.removeHint(firstHint.id, '$.authoring.parts[0].hints'))),
+      Hints.byPart(dispatch(model, Hints.removeOne(firstHint.id)), DEFAULT_PART_ID),
     ).toHaveLength(2);
   });
 });